enum valueType
