+++ src/selectors/_test/getSketchParams.test.js
@@ -1,8 +1,11 @@
-import test from 'tape'
 import deepFreeze from 'deep-freeze'
 import getSketchParams from '../getSketchParams'
 
-test('(engine) getSketchParams', function (t) {
+jest.mock('../../containers/ParamBar', () => null)
+jest.mock('../../containers/ParamCheckbox', () => null)
+jest.mock('../../containers/NodeSelect', () => null)
+
+test('(engine) getSketchParams', function () {
   let actual, expected
 
   const state = {
@@ -74,8 +77,8 @@ test('(engine) getSketchParams', function (t) {
   }
   actual = getSketchParams(state, 'xxx')
 
-  t.deepEqual(actual, expected,
-    'Returns key:value params for single sketch')
+  // Returns key:value params for single sketch
+  expect(actual).toEqual(expected)
 
   expected = {
     speedX: 0.3,
@@ -83,8 +86,8 @@ test('(engine) getSketchParams', function (t) {
   }
   actual = getSketchParams(state, 'yyy')
 
-  t.deepEqual(actual, expected,
-    'Returns key:value params for single sketch')
+  // Returns key:value params for single sketch
+  expect(actual).toEqual(expected)
 
   expected = {
     sketch_1: {
@@ -102,8 +105,8 @@ test('(engine) getSketchParams', function (t) {
 
   actual = getSketchParams(state)
 
-  t.deepEqual(actual, expected,
-    'Returns key:value params for all sketches if no id given')
+  // Returns key:value params for all sketches if no id given
+  expect(actual).toEqual(expected)
 
   expected = {
     sketch_1: {
@@ -118,13 +121,11 @@ test('(engine) getSketchParams', function (t) {
 
   actual = getSketchParams(state, null, 'aaa')
 
-  t.deepEqual(actual, expected,
-    'Returns key:value params for all sketches if scene ID given')
-
-  t.end()
+  // Returns key:value params for all sketches if no id given
+  expect(actual).toEqual(expected)
 })
 
-test('(engine) getSketchParams (min/max)', function (t) {
+test('(engine) getSketchParams (min/max)', function () {
   const state = {
     nodes: {
       '06': {
@@ -173,7 +174,6 @@ test('(engine) getSketchParams (min/max)', function (t) {
 
   actual = getSketchParams(state, '@@@')
 
-  t.deepEqual(actual, expected,
-    'Returns correct value based on min/max')
-  t.end()
+  // Returns correct value based on min/max
+  expect(actual).toEqual(expected)
 })
